wpt_babel/content-security-policy/style-src/style-src-injected-inline-style-allowed.html

<!DOCTYPE HTML>
<html>
 <head>
  <meta content="script-src 'self' 'unsafe-inline'; style-src 'unsafe-inline'" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("Injected inline style should load with 'unsafe-inline'");
document.addEventListener("securitypolicyviolation", t.unreached_func("Should not trigger a security policy violation"));
  </script>
 </head>
 <body>
  <div id="log">
  </div>
  <div id="content">
   Lorem ipsum
  </div>
  <script src="/content-security-policy/style-src/resources/style-src-inject-style.js">
  </script>
  <script>
   "use strict";

t.step(function () {
  onload = t.step_func_done(function (e) {
    var contentEl = document.getElementById("content");
    var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
    assert_equals(marginLeftVal, "2px");
    var marginRightVal = getComputedStyle(contentEl).getPropertyValue('margin-right');
    assert_equals(marginRightVal, "2px");
  });
});
  </script>
 </body>
</html>
